third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger/dynamic-script-tag-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that inline scripts and document.write scripts get different uiSourceCodes with different URLs.

Running: testOpenDevToolsAfterLoad
Both dynamic scripts were loaded.
Number of non-inline scripts with document url as a sourceURL: 0.

Running: testOpenDevToolsThenReload
Reloading page.
The first script with document url as a sourceURL to be seen is inline script.
Page reloaded.
